Conversion now tracks progress and can resume if restarted
The total run time (or perhaps some other measure of complexity) of
scripts seems to be limited by Google. This can result in the script
timing out if too many files are present in the folder to be converted.
To handle this gracefully, conversion is now performed into an "in
progress" export directory. If the script times out, it can be rerun,
and will attempt to resume conversion, skipping any documents that are
already present in the "in progress" export directory. Repeatedly
running the script should allow complete conversion, at which point the
export directory will be renamed with a timestamp as before.

The script will now generate "EXPORT-COMPLETE.txt" files in the
exported "subject" directories to allow resuming - these contain a
timestamp for when export of each "subject" directory was completed -
they can safely be deleted or ignored when export is complete.

A copyFolders function has been added which may be useful in backing
up/copying folders with the expected document structure.

Updated README.md